fix imports, placement of switchScreenKeepTtl
