+++ static/game/game.js
@@ -99,25 +99,6 @@ class Game {
         STATUS.textContent = won ? "CONGRATULATIONS!!! You won the game!"
             : "Sorry, you did not win. Break more wood next time.";
     }
-    static deleteGame() {
-        if (!Game.isMultiplayer) return;
-        const ttl = get("TTL");
-        const deadline = Game.rawMaterial.time * 1000 + 30 * 60 * 1000;
-        const interval = setInterval(() => {
-            const remaining = Math.max(0, Math.floor((deadline - Date.now()) / 1000));
-            const mins = Math.floor(remaining / 60);
-            const secs = remaining % 60;
-            ttl.textContent = `${mins}m${secs ? ` ${secs}s` : ""}`;
-            if (!remaining) {
-                clearInterval(interval);
-                document.body.querySelectorAll("div").forEach(div => {
-                    if (div.id != "GAME_DELETED") div.remove();
-                });
-                switchScreenKeepTtl("GAME_DELETED");
-                setTimeout(() => location.href = "/lounge/", 2000)
-            }
-        }, 1000);
-    }
 
     static world = WORLD;
     static player = 0;
